[test_nn] move init tests from test_nn
